Remove Cargo.lock from the repository
It is uncommon for library crates (such as this one) to check in their
Cargo.lock file [0]. Binaries (or other libraries) consuming this one ignore
it anyway, so all the file ends up being is a bit of a maintenance
burden, because making it makes it harder to identify truly outdated
dependencies according to our Cargo.toml specification. In addition, it
mostly just pollutes version control history.
Arguably, with the recent change to test the crate against our minimum
supported Rust version in CI, it now actually serves a real purposes in
that it prevents breakage if a dependency now requires a more recent
version of Rust. However, there is an argument to be made that in that
case our minimum version should align with that of the dependency
(because of the lock file being ignored; see above), but also that it is
basically inevitable for us to follow said bump anyway down the line.

For the sake of completeness, there is some open discussion with respect
the FAQ guidance referenced above [1]. However, it has not received much
traction and the arguments brought up are a bit lopsided and generally
not the strongest.

Weighing the pros and cons, this change removes the file from the
repository.

[0] https://doc.rust-lang.org/cargo/faq.html#why-do-binaries-have-cargolock-in-version-control-but-not-libraries
[1] https://github.com/rust-lang/cargo/issues/8728

Signed-off-by: Daniel Müller <deso@posteo.net>