Replace warning with semantic versioning notice
The warning in the README is arguably unnecessary and causing more
confusion than necessary. For one, it references [Semantic
versioning](https://semver.org/) but that differs from Cargo's treatment
of versions in subtle ways (in that it declares a free-for-all for
anything below 1.0).
Let's just state that we follow Cargo's semantic versioning rules and
leave it at that.

Signed-off-by: Daniel Müller <deso@posteo.net>