libbpf-cargo: Add support for struct_ops shadow objects
This change adds support for struct_ops shadow objects to the generated
skeletons. These allow for the pre-load modifications of fields in a
struct_ops object.
To make that happen we add a new generated struct_ops type that holds
pointers to all the known struct_ops objects and allows for modification
of their values before final load.

Signed-off-by: Daniel Müller <deso@posteo.net>