Adjust tcp_ca example to use struct_ops shadow data
This change adjusts the tcp_ca example to include the usage of
struct_ops shadow data. Specifically, we modify the name under which we
register the algorithm before attaching the program. We then make sure
that we can use the algorithm under the new name. In addition, we also
illustrate setting of a callback/program before load.

Signed-off-by: Daniel Müller <deso@posteo.net>