Refactor infrastructure for generating and loading BTF
The logic for generating BTF and loading BTF data is intermingled with
everything else, leading to shenanigans such as an entire Rust project
being created just to compile a C file -- for no good reason.
This change rips out the necessary bits and makes them reusable in the
form of the btf-gen crate.

Signed-off-by: Daniel Müller <deso@posteo.net>